# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AppCreate": {
        "additionalProperties": false,
        "properties": {
          "slug": {
            "description": "Host label of the app. Lower-case, no leading digit, no `--`.",
            "pattern": "^[a-z]([a-z0-9-]{0,38}[a-z0-9])?$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug"
        ],
        "title": "AppCreate",
        "type": "object"
      },
      "AppList": {
        "additionalProperties": false,
        "properties": {
          "apps": {
            "items": {
              "$ref": "#/components/schemas/AppSummary"
            },
            "title": "Apps",
            "type": "array"
          }
        },
        "required": [
          "apps"
        ],
        "title": "AppList",
        "type": "object"
      },
...
      "DeploymentCreate": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "default": "deploy",
            "enum": [
              "deploy",
              "rollback"
            ],
            "title": "Kind",
            "type": "string"
          },
          "release_id": {
            "pattern": "^rel_[a-z0-9]{20}$",
            "title": "Release Id",
            "type": "string"
          }
        },
        "required": [
          "release_id"
        ],
        "title": "DeploymentCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "One API for the command line, the GitHub Action, the console and the agent interface.\n\n* Every refusal is an RFC 9457 problem (`application/problem+json`) with a `code` from the error\n  catalogue and a `request_id`. The text is fixed; evidence stays in our logs under that id.\n* Every `POST` requires an `Idempotency-Key` header. A retry with the same key and body returns\n  the first answer with `Idempotency-Replayed: true`.\n* Edits to sharing rules require `If-Match` with the `ETag` from the last read.\n* A deployment is a long-running operation: `202` with a `Location` to poll.\n* Requests are rate-limited per credential; `429` carries `Retry-After`.\n",
    "title": "Small Software Cloud API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/apps": {
      "get": {
        "operationId": "list_apps_v1_apps_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHENTICATED`"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RATE_LIMITED`"
          }
        },
        "summary": "List Apps",
        "tags": [
          "v1"
        ]
      },
      "post": {
        "operationId": "create_app_v1_apps_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`IDEMPOTENCY_KEY_REQUIRED`"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHENTICATED`"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`IDEMPOTENCY_IN_FLIGHT`"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`IDEMPOTENCY_KEY_REUSED`, `VALIDATION_FAILED`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RATE_LIMITED`"
          }
        },
        "summary": "Create App",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/apps/{app_id}": {
      "get": {
        "operationId": "get_app_v1_apps__app_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z]{3}_[a-z0-9]{20}$",
              "title": "App Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHENTICATED`"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`NOT_FOUND`"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RATE_LIMITED`"
          }
        },
        "summary": "Get App",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/apps/{app_id}/environments/{environment_id}/deployments": {
      "post": {
        "description": "Start a deployment: 202 plus a ``Location`` to poll. Running it is the reconciler's job.",
        "operationId": "create_deployment_v1_apps__app_id__environments__environment_id__deployments_post",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z]{3}_[a-z0-9]{20}$",
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "environment_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z]{3}_[a-z0-9]{20}$",
              "title": "Environment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`IDEMPOTENCY_KEY_REQUIRED`"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHENTICATED`"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`NOT_FOUND`"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`DEPLOYMENT_IN_FLIGHT`"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`REFERENCE_NOT_FOUND`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RATE_LIMITED`"
          }
        },
        "summary": "Create Deployment",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/apps/{app_id}/environments/{environment_id}/grants": {
      "get": {
        "operationId": "get_grants_v1_apps__app_id__environments__environment_id__grants_get",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z]{3}_[a-z0-9]{20}$",
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "environment_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z]{3}_[a-z0-9]{20}$",
              "title": "Environment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHENTICATED`"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`NOT_FOUND`"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RATE_LIMITED`"
          }
        },
        "summary": "Get Grants",
        "tags": [
          "v1"
        ]
      },
      "put": {
        "description": "Replace the sharing rules of one environment. Requires ``If-Match`` with the current ETag.",
        "operationId": "put_grants_v1_apps__app_id__environments__environment_id__grants_put",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z]{3}_[a-z0-9]{20}$",
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "environment_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z]{3}_[a-z0-9]{20}$",
              "title": "Environment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHENTICATED`"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`FORBIDDEN`"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`NOT_FOUND`"
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`PRECONDITION_STALE`"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`REFERENCE_NOT_FOUND`, `VALIDATION_FAILED`"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`PRECONDITION_REQUIRED`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RATE_LIMITED`"
          }
        },
        "summary": "Put Grants",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/operations/{operation_id}": {
      "get": {
        "operationId": "get_operation_v1_operations__operation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z]{3}_[a-z0-9]{20}$",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`UNAUTHENTICATED`"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`NOT_FOUND`"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`VALIDATION_FAILED`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`RATE_LIMITED`"
          }
        },
        "summary": "Get Operation",
        "tags": [
          "v1"
        ]
      }
    },
...
  },
...
}